+++ src/platform/assets/components/AssetsListItem.vue
